优化PROOT_MODE逻辑  优化PROOT初始化  优化 linkerconfig/ld.config.txt 文件复制

+++ src/Main.java
@@ -57,7 +57,7 @@ public class Main {
 		return Build.VERSION.SDK_INT >= 26 ? ndkVersion + 2 : ndkVersion;
 	}
 
-	public static final String Version = "2.7.9";
+	public static final String Version = "2.8.0";
 
 	public static final String busyboxResourceName = "data/busybox";
 	public static final String ndkInstallShellResourceName = "data/ndk-install.sh";
@@ -563,10 +563,15 @@ public class Main {
 			return;
 		}
 
-		Main.PROOT_MODE = currentPackageContext.getApplicationInfo().targetSdkVersion > 28;
+		Main.PROOT_MODE = Build.VERSION.SDK_INT > Build.VERSION_CODES.P
+				&& currentPackageContext.getApplicationInfo().targetSdkVersion > Build.VERSION_CODES.P;
 
 		if (Main.PROOT_PATH == null) {
 			Main.PROOT_PATH = currentPackageContext.getApplicationInfo().nativeLibraryDir + "/libproot.so";
+		}
+
+		if (!Main.PROOT_MODE) {
+			return;
 		}
 
 		if (Main.PACKAGE_NAME_PATH == null) {
@@ -583,14 +588,14 @@ public class Main {
 		if (!ld_config_txt_file.exists() || ld_config_txt_file.length() == 0) {
 			try {
 
-				String inputPath = "/linkerconfig/ld.config.txt";
-
-				InputStream input = new FileInputStream(inputPath);
-				OutputStream output = new FileOutputStream(ld_config_txt_file);
-
-				IOUtils.streamTransfer(input, output, true);
-
-				ld_config_txt_file.setReadable(true, false);
+				String linkerconfigFilePath = "/linkerconfig/ld.config.txt";
+				File linkerconfigFile = new File(linkerconfigFilePath);
+				if (linkerconfigFile.exists()) {
+					InputStream input = new FileInputStream(linkerconfigFile);
+					OutputStream output = new FileOutputStream(ld_config_txt_file);
+					IOUtils.streamTransfer(input, output, true);
+					ld_config_txt_file.setReadable(true, false);
+				}
 			} catch (Throwable e) {
 				e.printStackTrace();
 			}
